Move tabelas de coeq e coaps para modulo DataTable
