--- src/features/acf/frontend/diabetes/modules/AcfPage/modules/PanelSelector/modules/List/modules/CoapsDataTable/model.ts
@@ -1,5 +1,0 @@
-import type { AppliedFiltersCoeq } from "../CoeqDataTable/model";
-
-export type AppliedFiltersCoaps = AppliedFiltersCoeq & {
-    careTeamName: Array<string>;
-};
