--- src/features/acf/frontend/diabetes/modules/AcfPage/modules/PanelSelector/modules/List/modules/CoeqDataTable/modules/DateRenderCell/modules/RenderDateTagCell.tsx
@@ -1,11 +1,0 @@
-import { TableTag } from "@/features/acf/frontend/common/TableTag";
-import { iconDetailsMap } from "@features/acf/frontend/common/TableTag";
-import React from "react";
-
-export const RenderDateTagCell: React.FC = () => (
-    <TableTag
-        theme="pending"
-        text="Não realizada"
-        icon={iconDetailsMap.pending}
-    />
-);
