--- src/features/acf/frontend/diabetes/modules/AcfPage/modules/PanelSelector/modules/List/modules/CoeqDataTable/modules/NameFormatter/index.tsx
@@ -1,10 +1,0 @@
-import { nameFormatter } from "../../../../../../logic"; //TODO: Checar este import
-
-type NameFormatterProps = {
-    value: string | null;
-};
-
-export const NameFormatter: React.FC<NameFormatterProps> = ({ value }) => {
-    if (!value) return <div data-testid="empty-return">{value}</div>;
-    return <div data-testid="name-return">{nameFormatter(value)}</div>;
-};
