+++ src/features/acf/frontend/diabetes/modules/AcfPage/modules/PanelSelector/modules/List/modules/DataTable/common/CpfOrBirthdayFormatter/__tests__/CpfOrBirthdayFormatter.test.tsx
