+++ src/features/acf/frontend/diabetes/modules/AcfPage/modules/PanelSelector/modules/List/modules/DataTable/index.tsx
@@ -1,0 +1,3 @@
+export { CoapsDataTable } from "./modules/CoapsDataTable";
+export { CoeqDataTable } from "./modules/CoeqDataTable";
+export type { AppliedFiltersCoaps, AppliedFiltersCoeq } from "./model";
