+++ src/features/acf/frontend/diabetes/modules/AcfPage/modules/PanelSelector/modules/List/modules/DataTable/modules/CoapsDataTable/consts.tsx
@@ -1,10 +1,10 @@
 import type { GridColDef } from "@mui/x-data-grid";
-import { CpfOrBirthdayFormatter } from "./modules/CpfOrBirthdayFormatter";
-import { DateRenderCell } from "./modules/DateRenderCell";
-import { NameFormatter } from "./modules/NameFormatter";
-import { RenderStatusTagCell } from "./modules/RenderStatusTagCell";
+import { NameFormatter } from "../../common/NameFormatter";
+import { CpfOrBirthdayFormatter } from "../../common/CpfOrBirthdayFormatter";
+import { DateRenderCell } from "../../common/DateRenderCell";
+import { RenderStatusTagCell } from "../../common/RenderStatusTagCell";
 
-export const coeqColumns: Array<GridColDef> = [
+export const coapsColumns: Array<GridColDef> = [
     {
         field: "patientName",
         headerName: "Nome",
@@ -115,6 +115,14 @@ export const coeqColumns: Array<GridColDef> = [
     {
         field: "communityHealthWorker",
         headerName: "Profissional responsável",
+        width: 230,
+        headerAlign: "left",
+        align: "left",
+        renderCell: NameFormatter,
+    },
+    {
+        field: "careTeamName",
+        headerName: "Nome da Equipe",
         width: 230,
         headerAlign: "left",
         align: "left",
